+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.2.0 - 2016-05-05
+### ADDED:
+- Add to wishlist event
+- View content event
+- Initiate checkout event
+- Complete registration event
+- Add to cart event
+
 ## 0.1.2 - 2016-03-15
 ### CHANGED:
 - Update readme file